combine the two cli tasks to one
